uyde-back/requirements.txt
django-environ
asgiref==3.8.1
Django==5.2
sqlparse==0.5.3
djangorestframework==3.15.1
django-extensions==3.2.3
django-registration==3.3
django-cors-headers==4.3.1
drf-yasg==1.21.7
drf-nested-routers
pillow==11.1.0


#postgres
psycopg2-binary==2.9.10

#redis
django-redis>=5.2.0

# AWS integration
boto3
django-storages
